Separate NavGraphs into different files for readability
